requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
defusedxml==0.7.1
idna==3.3
importlib-metadata==4.11.3
jira==3.1.1
keyring==23.5.0
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.1
Pillow==9.0.1
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jamf==0.7.3
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
tk==0.1.0
urllib3==1.26.9
zipp==3.7.0
